Remove openvswitch dependency for os-net-config
This removes the Requires for openvswitch.  os-net-config does not explicity
use openvswitch, but if ovs bridges or bonds are created it is needed.
Removing the dependency allows an image for ceph nodes to be created that
does not include openvswitch, and thus ovs bridges/bonds cannot be created
on these nodes.

Change-Id: I6a41693ffe310d873824d99bad3159cdd0350b20